commit_validator

+++ src/scripts/updateEvent.js
@@ -14,7 +14,7 @@ const updateEvent = (event) => {
     let end = new Date(endDate.value + 'T' + endTime.value)
     if (!duration(strat, end)) return;
     if (!errorDate(strat.getTime(), end.getTime())) return;
-
+    if (!checkForDelete(strat.getTime())) return;
 
     deleteAll()
